requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
prov==2.1.1
puremagic==1.29
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyMuPDF==1.26.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
redis==6.2.0
reportlab==4.4.1
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
scipy==1.16.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
traits==7.0.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1
